BUG: (12/17/09, GVG) Wrong function assigned to 40nm well resistors (Bug #2560) FIX: Function types changed. Amended code for resolution (Bug #2540)
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@10617 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/tool/user/tecEditWizard/General.java
@@ -53,7 +53,7 @@ public class General extends TechEditWizardPanel
 	{
 		TechEditWizardData data = wizard.getTechEditData();
 		stepSize.setText(Integer.toString(data.getStepSize()));
-        resolution.setText(Integer.toString(data.getResolution()));
+        resolution.setText(Double.toString(data.getResolution()));
         techName.setText(data.getTechName());
 		description.setText(data.getTechDescription());
 
@@ -69,7 +69,7 @@ public class General extends TechEditWizardPanel
 	{
 		TechEditWizardData data = wizard.getTechEditData();
 		data.setStepSize(TextUtils.atoi(stepSize.getText()));
-        data.setResolution(TextUtils.atoi(resolution.getText()));
+        data.setResolution(TextUtils.atof(resolution.getText()));
         data.setTechName(techName.getText());
 		data.setTechDescription(description.getText());
         data.setPSubstratelProcess(psubstrateButton.isSelected());
